search function in coorporated
- The urls are changed, or else cant use accomodations/search
- please check if the distance work once the geo_address function is done

+++ unihaven/models.py
@@ -26,7 +26,7 @@ class Accommodation(models.Model):
     available_until = models.DateField()
     beds = models.IntegerField()
     bedrooms = models.IntegerField()
-    rating = models.FloatField()
+    rating = models.IntegerField()
     daily_price = models.DecimalField(max_digits=10, decimal_places=2)
     owner = models.ForeignKey(PropertyOwner, on_delete=models.CASCADE, related_name="accommodations")
